Make the thread-id field be of the proper size on 64 bit machines.
This error was found by Gonzalo when he was looking at another problem
when thread IDs overflowed ints.

Currently we use gpointers in the io-layer for the threadid, but we were
only exposing the value as an int32.   The good news is that the code is
not really used right now, but could cause problems in the future as we
extend Process.

+++ mcs/class/System/System.Diagnostics/Process.cs
@@ -65,7 +65,7 @@ namespace System.Diagnostics {
 			 */
 			public IntPtr thread_handle;
 			public int pid; // Contains -GetLastError () on failure.
-			public int tid;
+			public IntPtr tid;
 			public string [] envKeys;
 			public string [] envValues;
 			public string UserName;
